+++ src/main.cpp
@@ -5,10 +5,13 @@
 #include <sstream>
 
 // Parthenon headers
+#include "bvals/boundary_conditions_generic.hpp"
+#include "defs.hpp"
 #include "globals.hpp"
 #include "parthenon_manager.hpp"
 
 // AthenaPK headers
+#include "bvals/boundary_conditions_apk.hpp"
 #include "hydro/hydro.hpp"
 #include "hydro/hydro_driver.hpp"
 #include "main.hpp"
@@ -89,6 +92,8 @@ int main(int argc, char *argv[]) {
     Hydro::ProblemInitPackageData = field_loop::ProblemInitPackageData;
   } else if (problem == "kh") {
     pman.app_input->MeshProblemGenerator = kh::ProblemGenerator;
+  } else if (problem == "lw_implode") {
+    pman.app_input->ProblemGenerator = lw_implode::ProblemGenerator;
   } else if (problem == "rand_blast") {
     pman.app_input->ProblemGenerator = rand_blast::ProblemGenerator;
     Hydro::ProblemInitPackageData = rand_blast::ProblemInitPackageData;
@@ -117,6 +122,23 @@ int main(int argc, char *argv[]) {
     msg << "Problem ID '" << problem << "' is not implemented yet.";
     PARTHENON_THROW(msg);
   }
+
+  const std::string REFLECTING = "reflecting";
+  using BF = parthenon::BoundaryFace;
+  using Hydro::BoundaryFunction::ReflectBC;
+  using parthenon::BoundaryFunction::BCSide;
+  pman.app_input->RegisterBoundaryCondition(BF::inner_x1, REFLECTING,
+                                            ReflectBC<X1DIR, BCSide::Inner>);
+  pman.app_input->RegisterBoundaryCondition(BF::outer_x1, REFLECTING,
+                                            ReflectBC<X1DIR, BCSide::Outer>);
+  pman.app_input->RegisterBoundaryCondition(BF::inner_x2, REFLECTING,
+                                            ReflectBC<X2DIR, BCSide::Inner>);
+  pman.app_input->RegisterBoundaryCondition(BF::outer_x2, REFLECTING,
+                                            ReflectBC<X2DIR, BCSide::Outer>);
+  pman.app_input->RegisterBoundaryCondition(BF::inner_x3, REFLECTING,
+                                            ReflectBC<X3DIR, BCSide::Inner>);
+  pman.app_input->RegisterBoundaryCondition(BF::outer_x3, REFLECTING,
+                                            ReflectBC<X3DIR, BCSide::Outer>);
 
   pman.ParthenonInitPackagesAndMesh();